consumer_dir/requirements.txt
kafka-python==2.0.2 
psycopg2-binary